services/tic-tac-toe/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.5.2,<3.0.0
nikitas-agents>=0.1.0
python-dotenv==1.0.0
